requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Flask-Session==0.5.0
Werkzeug==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
